Add regtest bitcoin test stand
Add test stand which is a small bitcoin network formed of three
bitcoin core nodes running in regtest mode connected to each
other. Stand is created using docker-compose, also, there is a
bunch of shell-scripts for managing it. A stand can be launched
and initialized with "./recreate-all.sh".
Stand is put to dir tools/regtest/.