Event storage: support postgres storage
Events can now be stored in postgres. Currently implementation
is straightforward, each event query results in a query to
database.
Event type is stored as string, event data is stored as JSONB.
Postgres datatype SERIAL is used for sequence numbers.
It should be noted that this can cause holes in sequence numbers
because SERIAL counter is never decremented, event in case of
transaction rollback, so failed transaction that executed insert
into table with events will cause hole to appear in seq.
SQL scripts to initialize database are put to tools/ dir.