HTTP API: add endpoint for listing txns
Add HTTP API endpoint "/get-transactions" which returns JSON array
of transactions in DB. It supports fitering: if a request has a
body, it should be a json object what may have fields "direction"
and "status". In case a field is present and has nonempty value,
only txns with equal corresponding value will be present in
resulting array.