Support manual withdraw confirmations
Withdrawals with amount more than a certain value, set in config,
are not processed immediately, but are held until they are
confirmed manually. Confirmation is done by request to '/confirm'
with transaction id. By default min value not needing confirmation
is 0, which means all withdrawals must be confirmed manually.

Fixes #1