Always send empty block hash on unconfirmed tx
Wallet tx events could in some cases have nonempty block hash set
even for txns with 0 confirmations. This could happen if event
about fresh unconfirmed tx was missed (for example because
processing was offline an was started only when tx already was
confirmed).
Fix that by always setting block hash to empty string in
notification in case number of confirmations is 0.