Integration test: forgotten wait in hotwallet test
Test TestCommonUsage/HotWallet/PersistsRestart stops and starts
processing (to check that it will remember hot wallet address
across restart), but the test did not wait for processing to
load (using 'env.waitForProcessing()'), so test could break
because processing would still be not ready to accept requests by
the time test starts interacting with it. Fix that.

See #4