Add method SendToMultipleAddresses to NodeAPI
This method uses RPC API 'sendmany' method to send money to
multiple addresses in one Bitcoin transaction. It's main use for
now is for testing (to simulate situation that client made such a
transaction), so this method does not support fee configuration -
fee is always auto-calculated and is paid by the sender.

See #11