Reliability: perform integrity check on start
Perform wallet and events integrity check in start. Currently this
means to check whether there was a withdraw and/or http
notification interrupted when processing was last stopped for
which processing can't determine whether that operation completed
normally or not. In case such a situation is detected, refuse to
start and print some info regarding what to do to fix it.

See #5