pgmitm: allow upstream pg connection failure
PgMITM now has an optional ability to allow upstream connection
failures - if this is enabled, upstream connection failure will
lead to log record and closure of client connection. If disabled,
proxy will, as before, panic.
This can be used for reliability test cases when postgres went
down (and proxy will therefore be unable to connect to it).

See #5