Support muting events by tx id
Add API for marking tx id as muted. HTTP callback sender in event
broker will skip events referring to such id when requesting
events from DB. Also, add support for this API to client lib and
to command-line client app.

Two modes are supported - giving exact tx uuid and special value
"current_problematic". The second form makes event broker check if
there is currently an event it it retrying to send, and, if there
is, take tx id from it (if there is one) and add that to
muted list.

Closes #14