Add internal withdraws without fee
In case a withdraw a requested which destination address belongs
to us (to one of accounts maintained by processing), it will now
be processed in a special way - it won't be sent to bitcoin node,
instead, fully confirmed outgoing and incoming txns will be
generated for it and all corresponding events will be created and
sent to backend. As a result, receiver will get the whole withdraw
amount, without fee subtracted.

Closes #15